pyproject.toml
[project]
name = "tau2-purple-agent"
version = "0.1.0"
description = "Advanced purple agent for the tau2-bench AgentBeats scenario."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk[http-server]>=0.3.20",
    "uvicorn>=0.38.0",
    "litellm>=1.55.0",
    "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.1",
]
